backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==0.27.0
pytesseract>=0.3.10
gunicorn>=21.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

